♻️ (server): Refactor code
