+++ apps/server/src/bookmarks/bookmarks.service.ts
@@ -12,7 +12,7 @@ import axios from 'axios';
 import { endpoints } from '@readable/common/constants';
 import * as FormData from 'form-data';
 import { TagsRepository } from '../tags/infrastructures/typeorm/repositories/tags.repository';
-import { InterestsRepository } from '@readable/interests/infrastructures/mongo/repositories/interest.repository';
+import { InterestsRepository } from '@readable/interests/infrastructures/typeorm/repositories/interest.repository';
 import cheerio from 'cheerio';
 
 @Injectable()
