+++ apps/server/src/interests/applications/usecases/get-my-interests-with-auth/get-my-interests-with-auth.usecase.ts
@@ -1,7 +1,7 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { Usecase } from '@readable/common/applications/usecase';
-import { Interest } from '@readable/interests/infrastructures/mongo/entities/interest.entity';
-import { InterestsRepository } from '@readable/interests/infrastructures/mongo/repositories/interest.repository';
+import { Interest } from '@readable/interests/infrastructures/typeorm/entities/interest.entity';
+import { InterestsRepository } from '@readable/interests/infrastructures/typeorm/repositories/interest.repository';
 import { User } from '@readable/users/domain/models/user.model';
 
 export class GetMyInterestsWithAuthUsecase implements Usecase<User, Interest[]> {
