+++ apps/server/src/interests/interests.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { InterestsRepository } from './infrastructures/mongo/repositories/interest.repository';
+import { InterestsRepository } from './infrastructures/typeorm/repositories/interest.repository';
 import { InterestsService } from './interests.service';
 import { InterestsResolver } from './interests.resolver';
 import { FindOrAddInterestWithAuthUseCase } from './applications/usecases/find-or-add-interest/find-or-add-interest-with-auth.usecase';
