+++ apps/server/src/user-bookmark/domain/model/user-bookmark.model.ts
@@ -1,5 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { CoreModel } from '@readable/common/models/core.model';
+import { Interest } from '@readable/interests/domain/interest.model';
 import { UrlInfo } from '@readable/url-info/domain/model/url-info.model';
 
 @ObjectType()
@@ -9,6 +10,9 @@ export class UserBookmarkBRFO extends CoreModel {
 
   @Field(type => UrlInfo)
   urlInfo: UrlInfo;
+
+  @Field(type => Interest)
+  interest: Interest;
 
   @Field(type => Date, { nullable: true })
   scheduledAt?: Date;
