+++ apps/server/src/user-bookmark/infrastructures/typeorm/entities/user-bookmark.entity.ts
@@ -1,6 +1,7 @@
 import { CoreEntity } from '@readable/common/infrastructures/typeorm/entities';
+import { Interest } from '@readable/interests/infrastructures/typeorm/entities/interest.entity';
 import { UrlInfo } from '@readable/url-info/infrastructures/typeorm/entities/url-info.entity';
-import { Column, Entity, ManyToOne } from 'typeorm';
+import { Column, Entity, JoinColumn, ManyToOne, OneToOne } from 'typeorm';
 
 @Entity('userBookmarks')
 export class UserBookmark extends CoreEntity {
@@ -9,6 +10,10 @@ export class UserBookmark extends CoreEntity {
 
   @ManyToOne(() => UrlInfo, urlInfo => urlInfo.userBookmarks)
   urlInfo: UrlInfo;
+
+  @OneToOne(() => Interest)
+  @JoinColumn()
+  interest: Interest;
 
   @Column({ type: 'datetime', nullable: true })
   scheduledAt?: Date;
